feat: ecran rendez vous fonctionnel. manque la verification des input avant les submits

+++ src/app/components/etudiants/etudiants.component.ts
@@ -45,10 +45,12 @@ export class EtudiantsComponent implements OnInit {
   }
 
   deleteEtudiant(ine : any){
-    this.etudiantService.delete(ine)
-    .subscribe(() => {
-      this.etudiants = this.etudiants.filter(etudiant => etudiant.ine != ine)
-    })
+    if(confirm("Are you sure to delete ")) {
+      this.etudiantService.delete(ine)
+      .subscribe(() => {
+        this.etudiants = this.etudiants.filter(etudiant => etudiant.ine != ine)
+      })
+    }
   }
 
   persistEtudiant(){
